store relocation
